committing the skeleton of threading stuff
