+++ p3/serverSNFS/Makefile
@@ -10,7 +10,7 @@ LDFLAGS +=  ../lib/libprotobuf.a ../lib/libprotobuf-c.a
 
 SERVER_EXEC = serverSNFS
 
-SRCS = server.c ../protobuf-model/ping.pb-c.c ../protobuf-model/fs.pb-c.c
+SRCS = filesystem.c server.c threading.c ../protobuf-model/ping.pb-c.c ../protobuf-model/fs.pb-c.c
 OBJS = ${SRCS:.c=.o}
 
 all: ${SERVER_EXEC}
