+++ p3/serverSNFS/server.h
@@ -1,0 +1,10 @@
+#pragma once
+
+struct _thread_args {
+  void * closure_data;
+  void * closure;
+  void * input;
+};
+
+typedef struct _thread_args thread_args;
+
